Adding a envelope clearance parameter for B0 ecal pacman disk
If the envelope attribute is set to true for pacman disk in the xml file, then the current parametrization may result in volume overlaps at the inner and outer radial extremes of modules with the envelope. This clearance parameter can be tuned to avoid that. It's a quick and dirty solution for now.