Defining envelope clearance parameter for the outer radial extremes
We only ensure that only the innermost corners of the individual modules fall within the specified bounds. When the envelope attribute is set to true, this will result in a overlap of envelope volume with modules on the outer radial extreme. Setting this clearance parameter should address this problem. It's set to a generous value in this case. But can be tuned on a case by case basis. Make sure that the envelope doesn't overlap with other volumes in the geometry.